Merge branch master from airbnb/css (#1)
* Advise against @extend, enable SCSS-lint ExtendDirective

@extend is unintuitive and dangerous, so we are recommending against its
use. As part of this commit, I enabled the SCSS-lint linter that will
warn against its use. I also revised content that mentioned @extend, to
maintain philosophical consistency.

* Add section about mixins

In my previous commit, I revised this document to advise against using
@extend, which included removing the section about mixins because it
only existed to recommend against using argumentless mixins in favor of
@extend. However, this document feels a bit incomplete without some
guidance on mixin usage, so I am adding a few sentences here.

* Prefer `border: 0` over `border: none`

Although the scss-lint configuration file here enforced `border: none`,
we chatted about it and agreed that we prefer `border: 0`.

* Default scss-lint severity to error

scss-lint 0.44.0 added a global severity configuration setting. Since
warnings are often ignored, we want to set this to error by default.

* Add section about using dash-cased for variable names

We want people to use a consistent style for variable names. We believe
that simple rules are most likely to be followed, and that consistency
is good, so we are settling on dash-cased, which is what we use for
everything else, such as classes, mixins, functions, and properties.

Along with this change, I am explicitly enabling the NameFormat
SCSS-Lint rule which enforces this style. Note that this is enabled in
SCSS-Lint by default, so this is functionally no different, but it is
good to be explicit about things.

* Enable PrivateNamingConvention scss-lint rule

Since we mention that prefixing names with underscores is an acceptable
way to denote that they are private, we want to enforce that they are
defined within the same file that they are used. The new
PrivateNamingConvention rule in scss-lint is designed for this purpose.

* Update README.md

Add Chinese(Simplified) Translation.